feat: switch to pytoml, update installs for multiprocess
