--- discord_bot/cli/__init__.py
@@ -1,3 +1,0 @@
-from discord_bot.cli.main import main
-
-__all__ = ['main']
