+++ docker/Dockerfile
@@ -31,21 +31,12 @@ RUN groupadd -g 1000 discord && \
 
 COPY discord_bot/ "${APPDIR}/discord_bot/"
 COPY alembic/ "${APPDIR}/alembic/"
-COPY requirements.base.txt "${APPDIR}/"
-COPY requirements.txt "${APPDIR}/"
-COPY requirements.sqlite.txt "${APPDIR}/"
-COPY requirements.postgres.txt "${APPDIR}/"
 COPY alembic.ini "${APPDIR}/"
-COPY setup.py "${APPDIR}/"
+COPY pyproject.toml VERSION "${APPDIR}/"
+COPY scripts/ "${WORKDIR}/scripts/"
 COPY docker/entrypoint.sh /entrypoint.sh
 
-RUN pip install "${APPDIR}"
-
-RUN if [ "$DB_TYPE" = "postgres" ]; then \
-        pip install -r "${APPDIR}/requirements.postgres.txt"; \
-    else \
-        pip install -r "${APPDIR}/requirements.sqlite.txt"; \
-    fi
+RUN pip install "${APPDIR}[all,${DB_TYPE}]"
 
 RUN chmod +x /entrypoint.sh
