+++ docker/docker-compose.multiprocess.yml
@@ -20,13 +20,13 @@ services:
   bot:
     build:
       context: ..
-      dockerfile: docker/Dockerfile
+      dockerfile: docker/Dockerfile.bot
     volumes:
       - ../volumes/cnf/discord.bot.cnf:/opt/discord/cnf/discord.cnf:ro
       - ../volumes/db:/opt/discord/db
       - ../volumes/bot-logs:/var/log/discord
     ports:
-      - "8080:8080"
+      - "8081:8080"
     depends_on:
       - redis
       - dispatcher
