--- pytest.ini
@@ -1,23 +1,0 @@
-[pytest]
-asyncio_mode = strict
-asyncio_default_fixture_loop_scope = function
-python_files = tests/*.py
-filterwarnings =
-    error
-    error::requests.exceptions.RequestsDependencyWarning
-    # pytest-freezegun uses the deprecated distutils LooseVersion
-    ignore::DeprecationWarning:pytest_freezegun
-    # discord.py imports audioop (deprecated in 3.11, removed in 3.13); message-based because
-    # audioop warns with stacklevel=2 so the warning is attributed to discord/player.py, not audioop
-    ignore:'audioop' is deprecated:DeprecationWarning
-    # discord.py calls asyncio.iscoroutinefunction() (deprecated in 3.14, removed in 3.16);
-    # message-based because the warning is attributed to discord/ext/commands/core.py, not asyncio
-    ignore:'asyncio.iscoroutinefunction' is deprecated:DeprecationWarning
-    # music.py uses TemporaryDirectory().name to get a path, immediately discarding the object
-    # (intentional pattern, see pylint:disable=consider-using-with comments); stdlib cleans them up
-    ignore:Implicitly cleaning up <TemporaryDirectory:ResourceWarning:tempfile
-    # pytest-asyncio leaks an unclosed event loop between tests in Python 3.11; the two AF_UNIX
-    # sockets below are the event loop's internal self-pipe wakeup mechanism and disappear with it;
-    # this does not reproduce on Python 3.12+ where event loop creation semantics changed
-    ignore:unclosed event loop:ResourceWarning
-    ignore:unclosed <socket.socket:ResourceWarning
