--- requirements.base.txt
@@ -1,17 +1,0 @@
-redis==7.4.0
-async-timeout==5.0.1
-discord.py==2.7.1
-pydantic==2.13.3
-pydantic-yaml==1.6.0
-opentelemetry-sdk==1.41.1
-# Dont pin these so dependabot updates are easier
-opentelemetry-instrumentation-requests
-opentelemetry-instrumentation-redis
-opentelemetry-instrumentation-logging
-opentelemetry-instrumentation-sqlalchemy
-opentelemetry-exporter-otlp
-click==8.3.3
-pyaml-env==1.2.2
-psutil==7.2.2
-pytz==2026.1.post1
-requests==2.33.1
