--- requirements.dispatcher.txt
@@ -1,1 +1,0 @@
--r requirements.base.txt
