--- requirements.postgres.txt
@@ -1,1 +1,0 @@
-asyncpg==0.31.0
