--- requirements.sqlite.txt
@@ -1,1 +1,0 @@
-aiosqlite==0.22.1
