--- requirements.txt
@@ -1,18 +1,0 @@
--r requirements.base.txt
-aiohttp>=3.9
-dappertable @ git+https://github.com/tnoff/dappertable.git@v1.0.0
-opentelemetry-instrumentation-sqlalchemy
-alembic==1.18.4
-ijson==3.5.0
-beautifulsoup4==4.14.3
-boto3==1.42.96
-google-api-python-client==2.194.0
-moviepy==2.2.1
-davey==0.1.5
-PyNaCl==1.6.2
-SQLAlchemy==2.0.49
-spotipy==2.26.0
-yt-dlp==2026.3.17
-# Dont pin, let yt-dlp install
-yt-dlp-ejs
-ytmusicapi==1.11.5
