--- setup.py
@@ -1,58 +1,0 @@
-import setuptools
-import os
-
-THIS_DIR = os.path.dirname(__file__)
-REQUIREMENTS_FILES = [os.path.join(THIS_DIR, 'requirements.txt')]
-VERSION_FILE = os.path.join(THIS_DIR, 'VERSION')
-
-for root, dirs, files in os.walk(os.path.join(THIS_DIR, 'cogs/plugins')):
-    for name in files:
-        if 'requirements.txt' in name.lower():
-            REQUIREMENTS_FILES.append(os.path.join(root, name))
-
-def read_requirements(file_path, _seen=None):
-    if _seen is None:
-        _seen = set()
-    if file_path in _seen or not os.path.exists(file_path):
-        return []
-    _seen.add(file_path)
-    base_dir = os.path.dirname(file_path)
-    reqs = []
-    with open(file_path) as f:
-        for line in f.read().splitlines():
-            if line.startswith('-r '):
-                included = os.path.join(base_dir, line[3:].strip())
-                reqs += read_requirements(included, _seen)
-            else:
-                reqs.append(line)
-    return reqs
-
-required = []
-for file_name in REQUIREMENTS_FILES:
-    # Not sure why but tox seems to miss the file here
-    # So add the check
-    required += read_requirements(file_name)
-
-# Try/catch mostly here for tox.ini
-try:
-    with open(VERSION_FILE) as r:
-        version = r.read().strip()
-except FileNotFoundError:
-    version = '0.0.1'
-
-setuptools.setup(
-    name='discord_bot',
-    description='Discord Bot',
-    author='Tyler D. North',
-    author_email='me@tyler-north.com',
-    install_requires=required,
-    entry_points={
-        'console_scripts' : [
-            'discord-bot = discord_bot.cli:main',
-            'discord-db-backup = discord_bot.cli.db_backup:backup_main',
-            'discord-db-restore = discord_bot.cli.db_backup:restore_main',
-        ]
-    },
-    packages=setuptools.find_packages(exclude=['tests', 'docs']),
-    version=version,
-)
