--- tests/requirements.txt
@@ -1,13 +1,0 @@
-fakeredis==2.35.1
-freezegun==1.5.5
-pylint-pydantic==0.4.1
-pylint==4.0.5
-pytest==9.0.3
-pytest-asyncio==1.3.0
-pytest-cov==7.1.0
-pytest-freezegun==0.4.2
-pytest-mock==3.15.1
-requests-mock==1.12.1
-# Currently required for pytest-freezegun, can remove once they have that fixed
-setuptools==82.0.1
-tox==4.53.0
