+++ tox.ini
@@ -7,11 +7,8 @@ env_list =
     py314
 
 [testenv]
-deps =
-    -r requirements.txt
-    -r requirements.sqlite.txt
-    -r tests/requirements.txt
-commands = 
+extras = all,test,sqlite
+commands =
     pylint discord_bot/
     pylint --rcfile .pylintrc.test  tests/
     pytest --cov=discord_bot --cov-fail-under=99 tests/
